+++ client/src/components/TicketManagement/TicketEditor.jsx
@@ -38,7 +38,6 @@ const TicketEditor = ({
     "underline",
     "strike",
     "list",
-    "bullet",
     "blockquote",
     "code-block",
     "color",